Add Command class hierarchy and CommandManager registry
Previous MessageReceivedEventListener class is replaced with
the CommandManager class to handle command invocation in a
more structured way and avoid long IF blocks. To achieve this
also an abstract class Command is created that represents a
command supported by the application. Each time a new command
needs to be added, a class that extends Command class should
be created and an instance of this newly created class needs
to be added to the command registry the CommandManager holds.

To avoid unathorized actions from users that lack specific
permissions the CommandManager implements an authorization
service that validates that a user executes a command while
having the required permissions. The authorization service
also validates that the bot has the required permissions on
the server it is executed.